Attempt to fix an issue where data items are not excluded properly.
We still need to implement depth on stringily and make sure request
info like cookies, query strings, post_data are being checked for
excluded items.

@Frank, I moved some things to functions as it takes it off the
prototype (shrinks code at the expense of uglier code). Also, I was
trying to see if we could share the stringify code for extra error
properties but we might need to revert that change.